chore: move client apis to a seperate module
