+++ crates/curp/src/client/connect.rs
@@ -1,0 +1,159 @@
+use std::ops::Deref;
+
+use async_trait::async_trait;
+use curp_external_api::cmd::Command;
+use parking_lot::RwLock;
+use tracing::debug;
+
+use crate::members::ServerId;
+use crate::rpc::ConfChange;
+use crate::rpc::FetchClusterResponse;
+use crate::rpc::Member;
+use crate::rpc::ProposeId;
+use crate::rpc::ReadState;
+use crate::tracker::Tracker;
+
+use super::ProposeResponse;
+
+/// `ClientApi`, a higher wrapper for `ConnectApi`, providing some methods for communicating to
+/// the whole curp cluster. Automatically discovery curp server to update it's quorum.
+#[async_trait]
+#[allow(clippy::module_name_repetitions)] // better than just Api
+pub trait ClientApi {
+    /// The client error
+    type Error;
+
+    /// The command type
+    type Cmd: Command;
+
+    /// Send propose to the whole cluster, `use_fast_path` set to `false` to fallback into ordered
+    /// requests (event the requests are commutative).
+    async fn propose(
+        &self,
+        cmd: &Self::Cmd,
+        token: Option<&String>, // TODO: Allow external custom interceptors, do not pass token in parameters
+        use_fast_path: bool,
+    ) -> Result<ProposeResponse<Self::Cmd>, Self::Error>;
+
+    /// Send propose configuration changes to the cluster
+    async fn propose_conf_change(
+        &self,
+        changes: Vec<ConfChange>,
+    ) -> Result<Vec<Member>, Self::Error>;
+
+    /// Send propose to shutdown cluster
+    async fn propose_shutdown(&self) -> Result<(), Self::Error>;
+
+    /// Send propose to publish a node id and name
+    async fn propose_publish(
+        &self,
+        node_id: ServerId,
+        node_name: String,
+        node_client_urls: Vec<String>,
+    ) -> Result<(), Self::Error>;
+
+    /// Send move leader request
+    async fn move_leader(&self, node_id: ServerId) -> Result<(), Self::Error>;
+
+    /// Send fetch read state from leader
+    async fn fetch_read_state(&self, cmd: &Self::Cmd) -> Result<ReadState, Self::Error>;
+
+    /// Send fetch cluster requests to all servers (That's because initially, we didn't
+    /// know who the leader is.)
+    ///
+    /// Note: The fetched cluster may still be outdated if `linearizable` is false
+    async fn fetch_cluster(&self, linearizable: bool) -> Result<FetchClusterResponse, Self::Error>;
+
+    /// Fetch leader id
+    #[inline]
+    async fn fetch_leader_id(&self, linearizable: bool) -> Result<ServerId, Self::Error> {
+        if linearizable {
+            let resp = self.fetch_cluster(true).await?;
+            return Ok(resp.leader_id.unwrap_or_else(|| {
+                unreachable!("linearizable fetch cluster should return a leader id")
+            }));
+        }
+        let resp = self.fetch_cluster(false).await?;
+        if let Some(id) = resp.leader_id {
+            return Ok(id);
+        }
+        debug!("no leader id in FetchClusterResponse, try to send linearizable request");
+        // fallback to linearizable fetch
+        self.fetch_leader_id(true).await
+    }
+}
+
+/// This trait override some unrepeatable methods in ClientApi, and a client with this trait will be able to retry.
+#[async_trait]
+pub(crate) trait RepeatableClientApi: ClientApi {
+    /// Generate a unique propose id during the retry process.
+    fn gen_propose_id(&self) -> Result<ProposeIdGuard<'_>, Self::Error>;
+
+    /// Send propose to the whole cluster, `use_fast_path` set to `false` to fallback into ordered
+    /// requests (event the requests are commutative).
+    async fn propose(
+        &self,
+        propose_id: ProposeId,
+        cmd: &Self::Cmd,
+        token: Option<&String>,
+        use_fast_path: bool,
+    ) -> Result<ProposeResponse<Self::Cmd>, Self::Error>;
+
+    /// Send propose configuration changes to the cluster
+    async fn propose_conf_change(
+        &self,
+        propose_id: ProposeId,
+        changes: Vec<ConfChange>,
+    ) -> Result<Vec<Member>, Self::Error>;
+
+    /// Send propose to shutdown cluster
+    async fn propose_shutdown(&self, id: ProposeId) -> Result<(), Self::Error>;
+
+    /// Send propose to publish a node id and name
+    async fn propose_publish(
+        &self,
+        propose_id: ProposeId,
+        node_id: ServerId,
+        node_name: String,
+        node_client_urls: Vec<String>,
+    ) -> Result<(), Self::Error>;
+}
+
+/// Update leader state
+#[async_trait]
+pub(crate) trait LeaderStateUpdate {
+    /// update
+    async fn update_leader(&self, leader_id: Option<ServerId>, term: u64) -> bool;
+}
+
+/// Propose id guard, used to ensure the sequence of propose id is recorded.
+pub(crate) struct ProposeIdGuard<'a> {
+    /// The propose id
+    propose_id: ProposeId,
+    /// The tracker
+    tracker: &'a RwLock<Tracker>,
+}
+
+impl Deref for ProposeIdGuard<'_> {
+    type Target = ProposeId;
+
+    fn deref(&self) -> &Self::Target {
+        &self.propose_id
+    }
+}
+
+impl<'a> ProposeIdGuard<'a> {
+    /// Create a new propose id guard
+    pub(crate) fn new(tracker: &'a RwLock<Tracker>, propose_id: ProposeId) -> Self {
+        Self {
+            propose_id,
+            tracker,
+        }
+    }
+}
+
+impl Drop for ProposeIdGuard<'_> {
+    fn drop(&mut self) {
+        let _ig = self.tracker.write().record(self.propose_id.1);
+    }
+}
